Fixed build issue
