+++ src/main/java/com/github/lukamon24/testsponge/TestSponge.java
@@ -1,0 +1,35 @@
+package com.github.lukamon24.testsponge;
+
+import com.google.inject.Inject;
+import org.apache.logging.log4j.Logger;
+import org.spongepowered.api.Server;
+import org.spongepowered.api.event.Listener;
+import org.spongepowered.api.event.lifecycle.ConstructPluginEvent;
+import org.spongepowered.api.event.lifecycle.StartingEngineEvent;
+import org.spongepowered.plugin.PluginContainer;
+import org.spongepowered.plugin.builtin.jvm.Plugin;
+
+@Plugin("testsponge")
+public class TestSponge {
+
+    public final PluginContainer container;
+    public final Logger logger;
+
+    @Inject
+    TestSponge(final PluginContainer container, final Logger logger) {
+        this.container = container;
+        this.logger = logger;
+    }
+
+    @Listener
+    public void onConstructPlugin(final ConstructPluginEvent event) {
+        // Perform any one-time setup
+        this.logger.info("Constructing testsponge");
+    }
+
+    @Listener
+    public void onServerStarting(final StartingEngineEvent<Server> event) {
+        // Any setup per-game instance. This can run multiple times when
+        // using the integrated (singleplayer) server.
+    }
+}
